🐛 Fix an issue where dungeon chests are empty
An issue has been occurring where dungeon chests are empty due to an attempt to populate "null" items into them. These "null" insertions fail, and throw an exception which bubbles up and interrupts both the loot generation and the dungeon generation.

For now, I've added a catch to consume this exception such that loot generation and dungeon generation are not interrupted, but the more insidious issue remains. It seems that these "null" item insertions are coming from merged loot settings from Dungeon Settings files that do not specify loot.

A few strategies to approach this might be to not merge empty loot settings, or to not obey loot rules that do not specify items or weights. Tbd.

Thanks to [@Shivaxi](https://twitter.com/Shivaxi) for bringing this to my attention by asserting that dungeons are empty. 😢 Sorry for your frustration, fam.