🐛 Fix issue where block brushes shared state
Previously, theme's block brushes were stateful, meaning that modifications to blockbrushes potentially affected future uses of the same block brush from the same theme, leading to some strange effects.

A simple example of this here was that the door block brush was used to provide the top and bottom to doors for delayed generation, only one or the other would be consumed, as specifying the "tops" or "bottoms" to doors required modifying the theme's door block brush. 🥴